put Slots in its own component..
