+++ src/components/component-playground/Control/Slots.astro
@@ -1,0 +1,5 @@
+<a-slots-control {...Astro.props}><slot /></a-slots-control>
+<script>
+import './Slots.css'
+import './Slots.ts'
+</script>
